requirements.txt
pinecone
boto3
python-dotenv
langchain
langchain-community
langchain-core
numpy
pandas 